+++ src/models/repositoryNode.ts
@@ -2,75 +2,50 @@ import * as vscode from 'vscode';
 import * as path from 'path';
 import { DockerAPIV2Helper } from '../utils/dockerUtils';
 import { TagNode } from './tagNode';
-import { TreeItemCollapsibleState } from 'vscode';
 import { RootNode } from './rootNode';
 
-
 export class RepositoryNode extends RootNode {
-
-    private _chldrenCount: number = 0;
-    public get chldrenCount(): number {
-        return this._chldrenCount;
-    }
+    public readonly iconPath = {
+        light: vscode.Uri.file(path.join(__filename, '..', '..', '..', 'resources', 'light', 'Repository_16x.svg')),
+        dark: vscode.Uri.file(path.join(__filename, '..', '..', '..', 'resources', 'dark', 'Repository_16x.svg'))
+    };
+    public readonly contextValue = 'repositoryNode';
+    private _childrenCount = 0;
 
     constructor(
-        public readonly label: string,
-        public readonly collapsibleState: vscode.TreeItemCollapsibleState,
-        private readonly dockerAPIV2Helper: DockerAPIV2Helper,
+        label: string,
+        private readonly client: DockerAPIV2Helper,
         onDidChangeTreeData: vscode.EventEmitter<vscode.TreeItem | undefined>,
-        parent: RootNode | undefined = undefined,
-        public readonly iconPath = {
-            light: vscode.Uri.file(path.join(__filename, '..', '..', '..', 'resources', 'light', 'Repository_16x.svg')),
-            dark: vscode.Uri.file(path.join(__filename, '..', '..', '..', 'resources', 'dark', 'Repository_16x.svg'))
-        }
+        parent: RootNode
     ) {
-        super(label, collapsibleState, onDidChangeTreeData, parent);
+        super(label, vscode.TreeItemCollapsibleState.Collapsed, onDidChangeTreeData, parent);
     }
 
-    getChildren(element?: RootNode | undefined): vscode.ProviderResult<vscode.TreeItem[]> {
-        return new Promise(async resolve => {
-            let chldrns: vscode.TreeItem[] = new Array<vscode.TreeItem>();
-
-
-            let resp = await this.dockerAPIV2Helper.getTags(this.label);
-
-            if (resp) {
-                resp.tags.forEach(tag => {
-                    chldrns.push(new TagNode(tag, tag, TreeItemCollapsibleState.Collapsed, this.label, this.dockerAPIV2Helper, this.onDidChangeTreeData, this));
-                });
-            }
-
-            (chldrns as TagNode[]).sort((a: TagNode, b: TagNode): number => {
-                if (a.tag === 'latest') {
-                    return -1;
-                }
-                if (b.tag === 'latest') {
-                    return 1;
-                }
-
-                if (!isNaN(+a.tag) && !isNaN(+b.tag)) {
-                    if (+a.tag < +b.tag) {
-                        return 1;
-                    }
-                    if (+a.tag > +b.tag) {
-                        return -1;
-                    }
-                } else {
-                    if (a.tag < b.tag) {
-                        return 1;
-                    }
-                    if (a.tag > b.tag) {
-                        return -1;
-                    }
-                }
-
-                return 0;
-            });
-
-            this._chldrenCount = chldrns.length;
-            resolve(chldrns);
-        });
+    get childrenCount(): number {
+        return this._childrenCount;
     }
 
-    contextValue = 'repositoryNode';
+    async getChildren(): Promise<TagNode[]> {
+        const response = await this.client.getTags(this.label as string);
+        const tags = response?.tags ?? [];
+        const nodes = tags.map(tag => new TagNode(tag, this.label as string, this.client, this.onDidChangeTreeData, this));
+        nodes.sort(compareTags);
+        this._childrenCount = nodes.length;
+        return nodes;
+    }
+}
+
+function compareTags(a: TagNode, b: TagNode): number {
+    if (a.tag === 'latest') {
+        return -1;
+    }
+    if (b.tag === 'latest') {
+        return 1;
+    }
+    const aNum = Number(a.tag);
+    const bNum = Number(b.tag);
+    if (!Number.isNaN(aNum) && !Number.isNaN(bNum)) {
+        return bNum - aNum;
+    }
+    return b.tag.localeCompare(a.tag);
 }